+++ partials/single/next-prev.php
@@ -19,16 +19,30 @@ if ( 'post' !== get_post_type() ) {
 $term_tax = get_theme_mod( 'ocean_single_post_next_prev_taxonomy', 'post_tag' );
 $term_tax = $term_tax ? $term_tax : 'post_tag';
 
-// Args.
-$args = array(
-	'prev_text'          => '<span class="title"><i class="fas fa-long-arrow-alt-left" aria-hidden="true"></i>' . esc_html__( 'Previous Post', 'oceanwp' ) . '</span><span class="post-title">%title</span>',
-	'next_text'          => '<span class="title"><i class="fas fa-long-arrow-alt-right" aria-hidden="true"></i>' . esc_html__( 'Next Post', 'oceanwp' ) . '</span><span class="post-title">%title</span>',
-	'in_same_term'       => true,
-	'taxonomy'           => $term_tax,
-	'screen_reader_text' => esc_html__( 'Read more articles', 'oceanwp' ),
-);
+// Vars.
+$prev_text = '<span class="title"><i class="fas fa-long-arrow-alt-left" aria-hidden="true"></i>'. esc_html__( 'Previous Post', 'oceanwp' ) .'</span><span class="post-title">%title</span>';
+$next_text = '<span class="title"><i class="fas fa-long-arrow-alt-right" aria-hidden="true"></i>'. esc_html__( 'Next Post', 'oceanwp' ) .'</span><span class="post-title">%title</span>';
+$screen_rt = esc_html__( 'Read more articles', 'oceanwp' );
 
 // Args.
+if ( $term_tax === 'pub-date' ) {
+	$args = array(
+		'prev_text'           => $prev_text,
+		'next_text'           => $next_text,
+		'in_same_term'        => false,
+		'screen_reader_text'  => $screen_rt,
+	);
+} else {
+	$args = array(
+		'prev_text'           => $prev_text,
+		'next_text'           => $next_text,
+		'in_same_term'        => true,
+		'taxonomy'            => $term_tax,
+		'screen_reader_text'  => $screen_rt,
+	);
+}
+
+// Display Next/Prev navigation.
 $args = apply_filters( 'ocean_single_post_next_prev_args', $args ); ?>
 
 <?php do_action( 'ocean_before_single_post_next_prev' ); ?>